updating no cycle
